feat: add standalone education page
